move the schemas to a subfolder
