+++ libs/newsletters-data-client/src/fixtures/newsletter-fixtures.ts
@@ -1,5 +1,5 @@
 import type { LegacyNewsletter } from '..';
-import type { NewsletterData } from '../lib/newsletter-data-type';
+import type { NewsletterData } from '../lib/schemas/newsletter-data-type';
 
 export const TECHSCAPE_IN_NEW_FORMAT: NewsletterData = {
 	identityName: 'tech-scape',
