+++ libs/newsletters-data-client/src/index.ts
@@ -1,14 +1,14 @@
 export * from './lib/api-response-type';
-export * from './lib/draft-newsletter-data-type';
+export * from './lib/schemas/draft-newsletter-data-type';
 export * from './lib/deriveNewsletterFields';
 export * from './lib/draft-service';
 export * from './lib/draft-storage';
 export * from './lib/draft-to-newsletter';
-export * from './lib/emailEmbedSchema';
+export * from './lib/schemas/emailEmbedSchema';
 export * from './lib/launch-service';
-export * from './lib/legacy-newsletter-type';
-export * from './lib/meta-data-type';
-export * from './lib/newsletter-data-type';
+export * from './lib/schemas/legacy-newsletter-type';
+export * from './lib/schemas/meta-data-type';
+export * from './lib/schemas/newsletter-data-type';
 export * from './lib/newsletter-storage';
 export * from './lib/storage-response-types';
 export * from './lib/transformDataToLegacyNewsletter';
