+++ libs/newsletters-data-client/src/lib/newsletter-storage/InMemoryNewsletterStorage.ts
@@ -1,11 +1,11 @@
-import type { DraftNewsletterDataWithMeta } from '../draft-newsletter-data-type';
-import { makeBlankMeta } from '../meta-data-type';
+import type { DraftNewsletterDataWithMeta } from '../schemas/draft-newsletter-data-type';
+import { makeBlankMeta } from '../schemas/meta-data-type';
 import type {
 	NewsletterData,
 	NewsletterDataWithMeta,
 	NewsletterDataWithoutMeta,
-} from '../newsletter-data-type';
-import { isNewsletterDataWithMeta } from '../newsletter-data-type';
+} from '../schemas/newsletter-data-type';
+import { isNewsletterDataWithMeta } from '../schemas/newsletter-data-type';
 import { StorageRequestFailureReason } from '../storage-response-types';
 import type {
 	SuccessfulStorageResponse,
