+++ libs/newsletters-data-client/src/lib/newsletter-storage/NewsletterStorage.ts
@@ -1,12 +1,16 @@
-import type { DraftNewsletterDataWithMeta } from '../draft-newsletter-data-type';
-import type { MetaData } from '../meta-data-type';
-import { createNewMeta, stripMeta, updateMeta } from '../meta-data-type';
+import type { DraftNewsletterDataWithMeta } from '../schemas/draft-newsletter-data-type';
+import type { MetaData } from '../schemas/meta-data-type';
+import {
+	createNewMeta,
+	stripMeta,
+	updateMeta,
+} from '../schemas/meta-data-type';
 import type {
 	NewsletterData,
 	NewsletterDataWithMeta,
 	NewsletterDataWithoutMeta,
-} from '../newsletter-data-type';
-import { isPartialNewsletterData } from '../newsletter-data-type';
+} from '../schemas/newsletter-data-type';
+import { isPartialNewsletterData } from '../schemas/newsletter-data-type';
 import type {
 	SuccessfulStorageResponse,
 	UnsuccessfulStorageResponse,
