+++ libs/newsletters-data-client/src/lib/newsletter-storage/s3-newsletter-storage.ts
@@ -1,12 +1,12 @@
 import type { S3Client } from '@aws-sdk/client-s3';
-import type { DraftNewsletterDataWithMeta } from '../draft-newsletter-data-type';
-import { makeBlankMeta } from '../meta-data-type';
+import type { DraftNewsletterDataWithMeta } from '../schemas/draft-newsletter-data-type';
+import { makeBlankMeta } from '../schemas/meta-data-type';
 import type {
 	NewsletterData,
 	NewsletterDataWithMeta,
 	NewsletterDataWithoutMeta,
-} from '../newsletter-data-type';
-import { isNewsletterDataWithMeta } from '../newsletter-data-type';
+} from '../schemas/newsletter-data-type';
+import { isNewsletterDataWithMeta } from '../schemas/newsletter-data-type';
 import type {
 	SuccessfulStorageResponse,
 	UnsuccessfulStorageResponse,
