+++ libs/newsletters-data-client/src/lib/schemas/emailEmbedSchema.ts
