+++ libs/newsletters-data-client/src/lib/transformWizardData.spec.ts
@@ -1,6 +1,6 @@
 import { TECHSCAPE_IN_NEW_FORMAT } from '../fixtures/newsletter-fixtures';
-import type { DraftNewsletterData } from './draft-newsletter-data-type';
-import { isDraftNewsletterData } from './draft-newsletter-data-type';
+import type { DraftNewsletterData } from './schemas/draft-newsletter-data-type';
+import { isDraftNewsletterData } from './schemas/draft-newsletter-data-type';
 import type { FormDataRecord } from './transformWizardData';
 import {
 	draftNewsletterDataToFormData,
